Class LTerm_widget.checkbutton

Checkbutton. A button that can be in active or inactive state.

inherit t

method label : string

The text displayed on the checkbutton.

method label_zed : Zed_string.t

The text displayed on the button.

method state : bool

The state of checkbutton; true means checked and false means unchecked.

method set_label : string -> unit

method on_click : ?switch:LTerm_widget_callbacks.switch -> (unit -> unit) -> unit

on_click ?switch f calls f when the button state is changed.