Class LTerm_widget.modal_frame

A widget displaying a frame around child widget. Unlike frame, the child widget is not expanded to take all available space; instead the child is centered and frame is drawn around it. This is a utility class for creation of modal dialogs and similar widgets.

inherit frame