Class LTerm_widget.radiobutton

Radiobutton. The button which implements radio object contract, so can be added to radiogroup.

inherit t

method state : bool

The state of the button; true if button is "on" and false if the button is "off".

method on : unit

Switches the button state to "on". Affects only how the button is drawn, does not change the state of the group the button is added to. Use radiogroup.switch_to instead.

method off : unit

Switches the button state to "off". Affects only how the button is drawn, does not change the state of the group the button is added to. Use radiogroup.switch_to instead.

method label : string

The text displayed on the radiobutton.

method label_zed : Zed_string.t

The text displayed on the button.

method set_label : string -> unit

method id : 'a

The id of the button.

method on_click : ?⁠switch:LTerm_widget_callbacks.switch -> (unit -> unit) -> unit

on_click ?switch f calls f when the button is clicked. You probably want to use radiogroup.on_state_change instead.