Class LTerm_widget.radiogroup

Radio group.

Radio group governs the set of radio objects. At each given moment of time only one of the objects in the "on" state and the rest are in the "off" state.

method on_state_change : ?switch:LTerm_widget_callbacks.switch -> ('a option -> unit) -> unit

on_state_change ?switch f calls f when the state of the group is changed.

method state : 'a option

The state of the group. Contains Some id with the id of "on" object in the group or None if no objects were added to the group yet.

method register_object : 'a radio -> unit

Adds radio object to the group

method switch_to : 'a -> unit

switch_to id switches radio group to the state Some id, calls radio.on method of the object with the given id and radio.off method of all other objects added to the group.