Class LTerm_widget.scrollable

Main object implementing scroll logic for coordination between a scrollable wigdet and a scrollbar widget.

scrollable_adjustment implements the main logic and provides a lowlevel interface for controlling how mouse events are translated to scroll offsets (set_mouse_mode) and the size of the scrollbar (set_scroll_bar_mode).

scrollable_document provides a higher level interface for configuring the operation of the scrollbar where the scrollbar is used to reflect the area of a page within a potentially larger document.

scrollbar_private is an internal interface between the scrollable object and a scrollbar used to exchange parameters needed to perform rendering.

inherit scrollable_adjustment

inherit scrollable_document

inherit scrollable_private