Class LTerm_read_line.read_password

Read-line engine for reading a password. The stylise method default to replacing all characters by '*'. You can also for example completely disable displaying the password by doing:

method stylise = ([||], 0)

Also showing completion is disabled.

inherit Zed_string.t engine

method eval : Zed_string.t

Returns the result as a UTF-8 encoded string.