Class LTerm_read_line.term

Class for read-line instances running in a terminal.

inherit 'a abstract

method run : 'a Lwt.t

Run this read-line instance.

method private exec : ?keys:LTerm_key.t list -> action list -> 'a loop_result Lwt.t

Executes a list of actions. Rememver to call Zed_macro.add self#macro action if you overload this method.

method editor_mode : LTerm_editor.mode React.signal

The current editor mode.

method set_editor_mode : LTerm_editor.mode -> unit

Set the current editor mode.

method vi_state : LTerm_vi.state

Get the current vi_state .

method bind : LTerm_key.t list -> action list -> unit

method draw_update : unit Lwt.t

Updates current display and put the cursor at current edition position.

method draw_success : unit Lwt.t

Draws after accepting current input.

method draw_failure : unit Lwt.t

Draws after an exception has been raised.

method prompt : prompt React.signal

The signal holding the prompt.

method set_prompt : prompt React.signal -> unit

Sets the prompt signal.

method size : LTerm_geom.size React.signal

The size of the terminal. This can be used for computing the prompt.

method key_sequence : LTerm_key.t list React.signal

The currently typed key sequence.

method completion_start : int React.signal

Index of the first displayed word in the completion bar.

method hide : unit Lwt.t

Hide this read-line instance. It remains invisible until show is called.

method show : unit Lwt.t

Show this read-line instance if it has been previously hidden.

val mutable visible : bool

Whether the instance is visible.

method create_temporary_file_for_external_editor : string

Create a temporary file and return its path. Used for editing input with an external command.

method external_editor : string

External editor command.