Class type LTerm_widget.box

Type of widgets displaying a list of widget.

inherit t

method add : ?⁠position:int -> ?⁠expand:bool -> t -> unit

add ?position ?expand widget adds a widget to the box. If expand is true (the default) then widget will occupy as much space as possible. If position is not specified then the widget is appended to the end of the widget list. It raises Out_of_range if the given position is negative or exceed the number of widgets.

method remove : t -> unit

remove widget remove a widget from the box.