Class type LTerm_widget.scrollable_private

Interface used by the scrollbar widget to configure the scrollbar and get parameters needed for rendering

method set_scroll_window_size : int -> unit

The attached scroll bar needs to provide its window size during set_allocation

method get_render_params : int * int * int

Provide the scroll bar with rendering parameters